+++ sentry-okhttp/src/main/java/io/sentry/okhttp/SentryOkHttpInterceptor.kt
@@ -116,6 +116,7 @@ public open class SentryOkHttpInterceptor(
               requestBuilder.removeHeader(BaggageHeader.BAGGAGE_HEADER)
               requestBuilder.addHeader(it.name, it.value)
             }
+            tracingHeaders.w3cTraceparentHeader?.let { requestBuilder.addHeader(it.name, it.value) }
           }
       }